// web/src/app/types/story.ts
export type allDataArrayType = [
  {
    id: number;
    age: number;
    gender: string;
    industry: string;
    username: string;
    detail: string;
    reason: string;
    employment: string;
    years: number;
    job_difficulty: number;
    job_struggles: string;
    job_hunt_difficulty: number;
    job_hunt_struggles: string;
    notebook: string;
    free: string;
    created_at: string;
  }
];

export type allDataType = 
  {
    id: number;
    age: number;
    gender: string;
    industry: string;
    username: string;
    detail: string;
    reason: string;
    employment: string;
    years: number;
    job_difficulty: number;
    job_struggles: string;
    job_hunt_difficulty: number;
    job_hunt_struggles: string;
    notebook: string;
    free: string;
    created_at: string;
  }